Keep the bidding-box <div> even if we're not displaying anything.
That way, when the browser reloads it at the end of the auction, we have someething to send (namely an empty div).

Also: Arguably make some unit tests clearer.  Pretty extensive overhaul

+++ project/app/models/player.py
@@ -146,7 +146,7 @@ class Player(models.Model):
             old_partner_pk = self.partner.pk
             Player.objects.filter(pk__in={self.pk, self.partner.pk}).update(partner=None)
 
-            if table is not None:
+            if table is not None and table.id is not None:
                 table.delete()
 
         self._send_partnershipt_messages(action=SPLIT, old_partner_pk=old_partner_pk)
